Split out logging crate
