+++ .github/workflows/pre-merge-rust.yml
@@ -27,6 +27,7 @@ on:
     - .github/workflows/pre-merge-rust.yml
     - 'lib/runtime/**'
     - 'lib/llm/**'
+    - 'lib/logging/**'
     - 'lib/bindings/**'
     - 'launch/**'
     - 'components/**'
@@ -38,7 +39,7 @@ jobs:
   pre-merge-rust:
     runs-on: ubuntu-latest
     strategy:
-      matrix: { dir: ['.', 'lib/bindings/python', 'lib/runtime/examples'] }
+      matrix: { dir: ['.', 'lib/bindings/python', 'lib/runtime/examples', 'lib/logging'] }
     permissions:
       contents: read
     steps:
