+++ lib/runtime/Cargo.toml
@@ -29,6 +29,9 @@ default = []
 integration = []
 
 [dependencies]
+# repo
+dynamo-logging = { workspace = true }
+
 # Use workspace dependencies where available
 anyhow = { workspace = true }
 async-nats = { workspace = true }
@@ -52,7 +55,6 @@ tokio = { workspace = true }
 tokio-stream = { workspace = true }
 tokio-util = { workspace = true }
 tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
 thiserror = { workspace = true }
 uuid = { workspace = true }
 url = { workspace = true }
