+++ lib/runtime/src/config.rs
@@ -142,40 +142,6 @@ impl RuntimeConfigBuilder {
     }
 }
 
-/// Check if an environment variable is truthy
-pub fn env_is_truthy(env: &str) -> bool {
-    match std::env::var(env) {
-        Ok(val) => is_truthy(val.as_str()),
-        Err(_) => false,
-    }
-}
-
-/// Check if a string is truthy
-/// This will be used to evaluate environment variables or any other subjective
-/// configuration parameters that can be set by the user that should be evaluated
-/// as a boolean value.
-pub fn is_truthy(val: &str) -> bool {
-    matches!(val.to_lowercase().as_str(), "1" | "true" | "on" | "yes")
-}
-
-/// Check whether JSONL logging enabled
-/// Set the `DYN_LOGGING_JSONL` environment variable a [`is_truthy`] value
-pub fn jsonl_logging_enabled() -> bool {
-    env_is_truthy("DYN_LOGGING_JSONL")
-}
-
-/// Check whether logging with ANSI terminal escape codes and colors is disabled.
-/// Set the `DYN_SDK_DISABLE_ANSI_LOGGING` environment variable a [`is_truthy`] value
-pub fn disable_ansi_logging() -> bool {
-    env_is_truthy("DYN_SDK_DISABLE_ANSI_LOGGING")
-}
-
-/// Check whether to use local timezone for logging timestamps (default is UTC)
-/// Set the `DYN_LOG_USE_LOCAL_TZ` environment variable to a [`is_truthy`] value
-pub fn use_local_timezone() -> bool {
-    env_is_truthy("DYN_LOG_USE_LOCAL_TZ")
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
